pyproject.toml
[tool.poetry]
name = "scrum-asistent"
version = "0.1.0"
description = ""
authors = ["JuanHurtadoPenta <160263212+JuanHurtadoPenta@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.41.1"
openai = "^1.58.1"
python-dotenv = "^1.0.1"
langchain = "^0.3.13"
requests = "^2.32.3"
langchain-openai = "^0.2.14"
python-pptx = "^1.0.2"
openpyxl = "^3.1.5"
streamlit-authenticator = "^0.4.1"
pymongo = "^4.10.1"
tabulate = "^0.9.0"
python-docx = "^1.1.2"
pypdf2 = "^3.0.1"
pyjwt = "^2.10.1"
streamlit-cookies-manager = "^0.2.0"
streamlit-antd-components = "^0.3.2"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
